+++ blocklets/image-bin/src/components/layout.jsx
@@ -6,13 +6,19 @@ import styled from '@emotion/styled';
 import Uploader, { UploaderProviderWrapper } from './uploader';
 import Exporter from './exporter';
 import { useUploadContext } from '../contexts/upload';
+import { useSessionContext } from '../contexts/session';
+
+const hasAdminPermission = (user) => ['admin', 'owner'].includes(user?.role);
 
 export default function Layout({ title }) {
   const { tab } = useUploadContext();
+  const { session } = useSessionContext();
 
   const addons = [];
   if (tab === 'bucket') {
-    addons.push(<Exporter key="exporter-addon" />);
+    if (hasAdminPermission(session?.user)) {
+      addons.push(<Exporter key="exporter-addon" />);
+    }
     addons.push(<Uploader key="uploader-addon" />);
   }
